# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaginatedResponse_TourBookingRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TourBookingRead"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "example": 20,
            "title": "Limit",
            "type": "integer"
          },
          "page": {
            "example": 1,
            "title": "Page",
            "type": "integer"
          },
          "pages": {
            "example": 5,
            "title": "Pages",
            "type": "integer"
          },
          "total": {
            "example": 100,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit",
          "pages"
        ],
        "title": "PaginatedResponse[TourBookingRead]",
        "type": "object"
      },
...
      "TourBookingCreate": {
        "properties": {
          "agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          },
          "amount": {
            "minimum": 0.0,
            "title": "Amount",
            "type": "number"
          },
          "listing_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listing Id"
          },
          "phone": {
            "minLength": 8,
            "title": "Phone",
            "type": "string"
          },
          "preferred_date": {
            "format": "date",
            "title": "Preferred Date",
            "type": "string"
          },
          "preferred_time": {
            "pattern": "^(morning|afternoon|evening)$",
            "title": "Preferred Time",
            "type": "string"
          },
          "student_name": {
            "minLength": 2,
            "title": "Student Name",
            "type": "string"
          },
          "tour_type": {
            "pattern": "^(specific_hostel|full_search)$",
            "title": "Tour Type",
            "type": "string"
          },
          "zone": {
            "title": "Zone",
            "type": "string"
          }
        },
        "required": [
          "student_name",
          "phone",
          "zone",
          "tour_type",
          "amount",
          "preferred_date",
          "preferred_time"
        ],
        "title": "TourBookingCreate",
        "type": "object"
      },
      "TourBookingRead": {
        "properties": {
          "agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          },
          "amount": {
            "title": "Amount",
            "type": "number"
          },
          "contacted": {
            "default": false,
            "title": "Contacted",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "linked_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linked User Id"
          },
          "listing_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listing Id"
          },
          "phone": {
            "title": "Phone",
            "type": "string"
          },
          "preferred_date": {
            "format": "date",
            "title": "Preferred Date",
            "type": "string"
          },
          "preferred_time": {
            "title": "Preferred Time",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "student_name": {
            "title": "Student Name",
            "type": "string"
          },
          "tour_type": {
            "title": "Tour Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "zone": {
            "title": "Zone",
            "type": "string"
          }
        },
        "required": [
          "id",
          "student_name",
          "phone",
          "zone",
          "tour_type",
          "amount",
          "preferred_date",
          "preferred_time",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "TourBookingRead",
        "type": "object"
      },
      "TourBookingUpdateStatus": {
        "properties": {
          "contacted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contacted"
          },
          "status": {
            "pattern": "^(pending_payment|confirmed|paid|completed|no_show|cancelled)$",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "TourBookingUpdateStatus",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Rumia Backend API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tours": {
      "get": {
        "description": "Fetch tour bookings. Agent sees own, Admin sees all.",
        "operationId": "list_bookings_api_v1_tours_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "description": "Page number (1-indexed)",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Page number (1-indexed)",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page (max 1000)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Items per page (max 1000)",
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_TourBookingRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tour Bookings",
        "tags": [
          "Tours"
        ]
      },
      "post": {
        "description": "Book a hostel tour. Public (optional auth).",
        "operationId": "create_booking_api_v1_tours_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TourBookingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourBookingRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Tour Booking",
        "tags": [
          "Tours"
        ]
      }
    },
    "/api/v1/tours/me": {
      "get": {
        "description": "Fetch tour bookings linked to current student user. Authenticated.",
        "operationId": "list_my_tours_api_v1_tours_me_get",
        "parameters": [
          {
            "description": "Page number (1-indexed)",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Page number (1-indexed)",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page (max 1000)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Items per page (max 1000)",
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_TourBookingRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List My Tour Bookings",
        "tags": [
          "Tours"
        ]
      }
    },
    "/api/v1/tours/{booking_id}": {
      "get": {
        "description": "Fetch single booking by ID. Authorized user/agent/admin.",
        "operationId": "get_booking_api_v1_tours__booking_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "title": "Booking Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourBookingRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Booking Details",
        "tags": [
          "Tours"
        ]
      }
    },
    "/api/v1/tours/{booking_id}/status": {
      "patch": {
        "description": "Update booking status or contacted flag. Agent or Admin.",
        "operationId": "update_booking_status_api_v1_tours__booking_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "title": "Booking Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TourBookingUpdateStatus"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourBookingRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Booking Status",
        "tags": [
          "Tours"
        ]
      }
    },
...
  }
}